long long _data now long data
Was causing overflow issues when & the fault bits above 712 degrees.

+++ MAX31855_DEG_C.h
@@ -29,7 +29,7 @@ class MAX31855_DEG_C
  private:
     void ReadSPI(void);
     
-    long long _data;
+    long  _data;
     long _extractedDataH;
     long _extractedDataC;
